fix auth
